SAWScript: add an "include_once".
This is the same as "include" except that it skips any file that's
already been included.

This version does not interact with the REPL very well: the include
history does not persist past a single REPL input, and doesn't extend
to nested subshells. I'm going to hold off on improving that until
later, as currently carrying around the state for it is problematic,
or at least quite messy. In the not-too-distant future we should have
more infrastructure supporting incremental execution for the REPL,
and it'll be easier to add additional bits like this.