SAWScript: Don't timestamp prints by default.
If you really want the timestamping behavior, you can pass -T
on the command line to get it back.

For more serious timing there's the time and with_time SAWScript
builtins.

Remove the logic for pruning the timestamps from test-and-diff.sh.

Update all the test reference outputs, because unfortunately the
pruning had not been removing the space after the timestamp itself.
This changeset is therefore best read with git diff -w. :-|