Naming reform for the saw-core prettyprinter code: part 2.
The convention is that functions named 'pp' produce strings, functions
named 'pretty' produce docs. Adjust the external interface
accordingly.

To wit:

First, take scShowTerm from SAWCore.SharedTerm:

 - Rename it to ppTerm, because it's the preferred interface for
printing terms. (It uses the printing environment from the
SharedContext; not having that makes the output a lot more verbose.)

 - Add a corresponding prettyTerm that produces a Doc.

 - Move these to SAWCore.Term.Certified so they can be used for
   printing errors there. (They can't live in SAWCore.Term.Pretty
   because that doesn't have the SharedContext, and leaving them
   in SharedTerm would create an import cycle.)

Then rename the external interface of SAWCore.Term.Pretty as
follows:

 - First, the printing functions that can be used from pure code and
   use an empty naming environment:

      ppTerm -> prettyTermPure
      scPrettyTerm -> ppTermPure
      showTerm -> ppTermPureDefaults

   The last implicitly uses SAWSupport.Pretty.defaultOpts. (Ideally we
   shouldn't have any code that doesn't honor the user's printing
   options, but that's a long way off.)

 - Second, there are four interfaces that use explicit naming
   environments:

      ppTermInCtx -> prettyTermWithNameList
      ppTermWithNames -> prettyTermWithEnv
      showTermWithNames -> ppTermWithEnv
      ppTermContainerWithNames -> prettyTermContainerWithEnv

Update all the call sites.

I decided I did not want to make two passes through the call sites,
and so I've aggressively fixed as many of them as possible to use
either ppTerm or prettyTerm rather than the pure versions. This
fixes any number of ugly prints and closes #2802.

It is also a large part of #1277, and relevant to #747.