More prettyprinting name reform, higher up in the system.
pretty* returns a Doc, pp* returns text.

There are now two versions of the prettyprinting for TypedTerm, one
that's pure and one that takes a SharedContext and a PPS.Opts and
returns in IO. The latter's preferred because it allows printing
SAWCore Terms with the preferred Term-printing interface. However,
a lot of the users aren't ready to do that.

This changes more prints over to the preferred method, although
perhaps not many that are all that readily accessible.