Use bitvectors for the array indexes, when generating uninterpreted functions.
To do this, we first count how many symbols we need for each type, and then
we pick a bit-vector type that can represent all of these.

This should help for solvers, such as bitwuzla, which do not support integers.
It also helps in the special case where a function only has one uninterpred results, in which case we don't crate an array at all.