SAWScript: Add exception frobs for the new error stuff.
This is all rather ad hoc, and some of it resembles chewing gum rather
than logic; however, that's because the exception handling stuff in
here is already fundamentally chewing-gum-based.

The exception thrown by the new errors ("Fatal") doesn't have
meaningful error text in it, because the error text has already been
printed. (In the remote API's version, it'll be collected into monad
state. That doesn't exist yet; doing it as desired depends on
disentangling the remote API from the TopLevel monad TopLevel; we're
not there yet though we're getting closer.)

Therefore the upper-level places that catch and print exceptions need
an extra case for Fatal to avoid printing an extra pointless output
line. This also applies to the "fails" command.