Make the prettyprinter options available from the SharedContext.
Move them to an IORef; stick one copy in the SharedContext and another
in the TopLevelRO. Allow fetching them from the SharedContext outside
of saw-core, even though this isn't really the mission of the
SharedContext.

This drastically reduces the number of places they have to be passed
around explicitly.

Avoid use of PPS.defaultOpts in various places where the proper value
is now available, less by specific intent than as a result of
incremental changes. Hunting the rest down will happen at some later
point.