MIR: Adapt to TyCoroutine gaining more fields
This bumps the `mir-json` and `crucible` submodules to bring in the following
changes:

* https://github.com/GaloisInc/mir-json/pull/214
* https://github.com/GaloisInc/crucible/pull/1686

The latter PR updates the definition of the `TyCoroutine` data constructor,
which requires some corresponding code changes in `saw-central`. The latter PR
also bumps the `mir-json` schema version to `8`. I will regenerate the
vendored-in MIR JSON files in the SAW test suite in a follow-up commit.